myproject/requirements.txt
Django==5.1.7
google-generativeai==0.8.5
Pillow==11.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
